Latest updates as of 10/04/2020 (#11)
* &mut self instead of mut self for MessageBuilder methods

* http: make error module public

The `http::error::Error` type is re-exported from `http` as `HttpError`,
and while the `ErrorResponse` and `DiscordJsonError` types are public,
they aren't re-exported from the parent module and the `http::error`
module is private.

This means that the `Error::UnsuccessfulRequest` variant ends up not
being as useful as it could be, as seen here:
<https://docs.rs/serenity/0.8.0/serenity/http/enum.HttpError.html#variant.UnsuccessfulRequest>

Since these 3 types are the only ones in the `http::error` module, this
commit just makes the module public.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>

* Implement `Color` type alias to `Colour` (#791)

* Reexport `Timestamp` from the builder's module.

While it's legal to use it in the public API through generics, it's
impossible to construct this yourself. You cannot import it, nor even
view its respective docs, because it's private to `CreateEmbed`'s
module. This commit lifts these restrictions.

* Add impls for borrowed `Arc`s to `CacheHttp`

This should allow for passing `&ctx.http` to a function expecting a type
that implements `CacheHttp`.

* Only set the content type header if there's a body

* Store command names in lowercase when case-insensitivity is on

* Implement various `kick_with_reason()` methods. (#797)

* Update changelog for 0.8.1

* Bump command_attr's version

* Bump version

* Update depedency version of commmand_attr

* Always send the `Content-Length` header (#819)

* Fix missing async specifier, update crate versions

Co-authored-by: MaxOhn <ohn.m@hotmail.de>
Co-authored-by: Vivian Hellyer <vivian@hellyer.dev>
Co-authored-by: Elinvynia <59487684+Elinvynia@users.noreply.github.com>
Co-authored-by: Alex M. M <acdenissk69@gmail.com>
Co-authored-by: nitsuga5124 <nitsuga5124@gmail.com>
Co-authored-by: Lakelezz <12222135+Lakelezz@users.noreply.github.com>